Finished the assignment
First I added all the gulfiles /linters,           Then using the files I made sure that the all files (html, css, and java) were synced to the syntax. Next i went to the package.json file to apply the versions of each package.
Made the changes to the page using the css file.                                                  Lastly I went into the main.js file to work on the drop down menu and the modal.